fix(blorktools): label loading
